The two original files, untouched since 2011
After not touching this code for 10 years, I was pleased to find that it still mostly worked on my machine. I was able to play a full game with no errors, but it started giving errors after my first game finished and I clicked "New Game."

I'm looking forward to getting the program up-to-date and error-free in my spare time, and ultimately enabling it to be played in-browser to make it easily accessible. Eventually, it would be cool to add custom game modes that use the same board but different rules (i.e. "Connect Three", 3+ colors of chip, 2+ chips per turn, etc.).

If you have any questions or ideas, please let me know! Also, forgive any sloppiness if you can - I wrote this as a wee lad many moons ago, and I still had a lot to learn.